+++ concordium-node/src/common/serialization/serializable.rs
@@ -185,7 +185,7 @@ impl Serializable for UCursor {
         let mut self_from = self.sub(self.position())?;
         let self_from_len = self_from.len();
 
-        self_from_len.serialize(archive)?;
+        (self_from_len as u32).serialize(archive)?;
         std::io::copy(&mut self_from, archive)?;
 
         Ok(())
